More css and refactoring
